Split model picker surface tests
